changed member links to link values
